backend/requirements.txt
flask
youtube-search
ftfy
duckduckgo-images-api
firebase-admin
requests
python-dotenv